(trivial) move StorageEncryptionVersion and StorageReadWriteError to stored_dict.py
