+++ electrum/storage.py
@@ -28,7 +28,6 @@ import stat
 import hashlib
 import base64
 import zlib
-from enum import IntEnum
 from typing import Optional
 
 import electrum_ecc as ecc
@@ -37,7 +36,6 @@ from . import crypto
 from .util import (profiler, InvalidPassword, WalletFileException, bfh, standardize_path,
                    test_read_write_permissions, os_chmod)
 
-from .wallet_db import WalletDB
 from .logging import Logger
 
 
@@ -47,13 +45,9 @@ def get_derivation_used_for_hw_device_encryption():
             "/1112098098'")  # ascii 'BIE2' as decimal
 
 
-class StorageEncryptionVersion(IntEnum):
-    PLAINTEXT = 0
-    USER_PASSWORD = 1
-    XPUB_PASSWORD = 2
+from .stored_dict import StorageEncryptionVersion, StorageReadWriteError
 
 
-class StorageReadWriteError(Exception): pass
 
 
 class StorageOnDiskUnexpectedlyChanged(Exception): pass
